+++ CPU_Renderer/CPU_Renderer.vcxproj.filters
@@ -24,12 +24,18 @@
     <ClCompile Include="Source\Colors\Colors.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Source\Input\Input.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Source\Colors\Colors.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Source\Display\Display.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\Input\Input.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
