+++ CPU_Renderer/Source/Display/Display.cpp
@@ -103,3 +103,10 @@ void draw_rectangle(int x, int y, int w, int h, uint32_t* color) {
 		}
 	}
 }
+
+void destroy() {
+	free(color_buffer);
+	SDL_DestroyRenderer(renderer);
+	SDL_DestroyWindow(window);
+	SDL_Quit();
+}
