+++ CPU_Renderer/Source/Input/Input.h
@@ -1,0 +1,13 @@
+#ifndef INPUT_H
+#define INPUT_H
+
+#include "SDL.h"
+
+void keyboard_key_down(SDL_KeyboardEvent key);
+void keyboard_key_up(SDL_KeyboardEvent key);
+void mouse_button_down(SDL_MouseButtonEvent button);
+void mouse_button_up(SDL_MouseButtonEvent button);
+void mouse_motion(SDL_MouseMotionEvent motion);
+void mouse_wheel(SDL_MouseWheelEvent wheel);
+
+#endif
